Use SafeDumper to dump logs to YAML strings
Also add some additional representers to the dumper, as we found that
the following tags are in the logs, which cause exceptions when using
SafeDumper:
  !!python/name:cros.factory.gooftool.gooftool.CreateReportArchiveBlob
  !!python/object/apply:collections.OrderedDict
  !!python/object:autotest_lib.client.cros.rf.agilent_scpi.Traces
  !!python/object/new:cros.factory.hwid.ProbedComponentResult
  !!python/object/new:dbus.String

The OrderDict, ProbedComponentResult, and String are subclasses of
native classes. We adds the subclasses representer.

The CreateReportArchiveBlob, Traces, and other general objects are
outputed their __dict__ attributes.

BUG=chrome-os-partner:19652
TEST=make test

Change-Id: Ie8c6381e7a8712745a631edaff77a9459ec85f47
Reviewed-on: https://gerrit.chromium.org/gerrit/57906
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>