suspend_resume: swap powerd_suspend thread with wake extension
To ensure the test fails immediately when a powerd_suspend call fails we
can put the powerd_suspend into the main thread, and move the wakealarm
managing code to the thread.

This also refactors the thread from a daemon thread to a regular thread that
starts with the test sequence and is joined by the tearDown().

BUG=chrome-os-partner:19917
TEST=Manual, run powerd_suspend on DUT.
Verify test fails immediately when powerd_suspend returns failure by typing
during the test (should return 2 for wake event during suspend).

Change-Id: I2b01d7542b6aa3fc3d67e343ec3e72102f1a64f3
Reviewed-on: https://gerrit.chromium.org/gerrit/58949
Reviewed-by: Jon Salz <jsalz@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>