E5071C mock: add httpd for screenshot requests
There is a mode in DUT request screenshots via http protocol.
To fully enable the mock with existing flow, we have to simulate this
http server as well. A command line option is added for this. In
addition, the port is now configurable via command line.

BUG=chrome-os-partner:20240
TEST=sudo python ./e5071c_mock.py --httpd
Manually execute VSWR test with screenshot enabled.

Change-Id: I452689e29fad5ba71caedc64414836cf5bad54e5
Reviewed-on: https://gerrit.chromium.org/gerrit/61056
Commit-Queue: Chun-Ta Lin <itspeter@google.com>
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Tested-by: Chun-Ta Lin <itspeter@google.com>