RMA: add a GetDeviceInfo function to the simple rma shopfloor
To support HWIDv3 generation on the DUT, it will need to be able to get some
basic information about what it is, particularly regarding aspects which are
not probe-able. This adds the GetDeviceInfo function to collect such
information from the DUT's RMA YAML file.

This will eventually work together with HWIDv3 decoding in the rma shopfloor
to put the fields in question into the YAML file to begin with, as well as
RMA center sourced events for replacement parts.

BUG=chrome-os-partner:20537
TEST=Manual, run shopfloor, request GetDeviceInfo from DUT.

Change-Id: I7cdcce78515e389a6ff709404cc95049288c561a
Reviewed-on: https://gerrit.chromium.org/gerrit/61230
Reviewed-by: Drew Haven <haven@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>