RMA: Add in a component name translation capability for RMA shopfloor
In many cases the component name extracted from a HWID does not match what a
given factory or RMA center uses to name the same component, so we need
some way to translate between the names. This adds in an optional config
value from the rma_config_board.yaml file to define the translation.

BUG=chrome-os-partner:20537
TEST=Manual, run shopfloor server, check translation output.

Change-Id: I97b9869bbfaa54709f79bcc57a50ea87775561da
Reviewed-on: https://gerrit.chromium.org/gerrit/61986
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>