Suspend_resume: give up wakealarm extension upon write failure
If the test happens to suspend at just the right(wrong) time, there is
a race where the write to the wakealarm could fail if the timer had
already expired (e.g. we already woke up). In general the failure to
write to the wakealarm should not fail silently anyway.

BUG=chrome-os-partner:19958
TEST=Manual, run suspend_resume test.

Change-Id: I2711a8fdc0715a076767a74b5b7137af10cb9ee5
Reviewed-on: https://gerrit.chromium.org/gerrit/62049
Reviewed-by: Jon Salz <jsalz@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>