Minijack: Change border-collapse to separate for dataTable
Currently, when there are many columns in a dataTable in minijack
frontend, the header and footer of the table would be misaligned.
This change fix the problem.

BUG=None
TEST=Manual
Start minijack frontend, check url /build, header and footer would be
aligned even when all columns are shown.

Change-Id: I761f9ff978e4415f8881edf467b590eacc2c9492
Reviewed-on: https://gerrit.chromium.org/gerrit/62807
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>