Make finalize_bundle leverage the new-style active test_list mechanism
The old active test_list method used a symlink to the active test_list in
/usr/local/factory/test_lists/active the new-style uses a file containing
the test_list extension in /usr/local/factory/py/test/test_lists/ACTIVE. We
need to use the new-style if we want to change the default test_list in
finalize_bundle.

To support the v1 style test_lists we also check if the test_lists.py file
exists and if so use the old symlink method.

BUG=chrome-os-partner:21209
TEST=Manual, build a bundle with finalize_bundle using a different test_list
and verify that test_list is used.

Change-Id: I773214388680d551a6921daaa6312f2ba935361f
Reviewed-on: https://gerrit.chromium.org/gerrit/62870
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>