Minijack: Fix bug when number of component_class isn't 27.
The number was incorrectly hard-coded in javascript file, change the way
to handle this.
Also remove select menu when all value of the column are the same.
Also change the ordering of devices in hwid_view to match the one used
in tests_view.

BUG=None
TEST=Manual
Start minijack frontend, view /test when number of component_class in
database is not 27, there shouldn't be any javascript error in browser's
console.

Change-Id: I21a11e74054772f7cb21193e7fad468e2fe52297
Reviewed-on: https://gerrit.chromium.org/gerrit/63184
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>