Add an auto-update option to miniomaha.py
If auto-update option is enabled, miniomaha engine will periodically
check if the current images are up-to-date.

Currently the update check will be triggered in two cases. The first case is
when the server is idled for a certain period. The second case is when
there is no update check for a long period.

Note that in the later case, there may be file inconsistency issues if
the update check is forced between the update procedure of some machines.

BUG=chrome-os-partner:20735
TEST=Singal machine manual test

Change-Id: I857a156ad26cec1239f19fd9df93e462bedb5d2c
Reviewed-on: https://gerrit.chromium.org/gerrit/62082
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Chun-Yen Ho <chho@google.com>
Tested-by: Chun-Yen Ho <chho@google.com>