Make x and y segment count values arguments in touchpad.py
In some cases we may want to vary the level of complexity of the touchpad
test by adjusting the segments we have to be touched in the test. This makes
the default the same at 5 but allows the test_list to adjust the value.

BUG=None.
TEST=Manual, run with a test_list entry using
dargs={'x_segments': 3,
       'y_segments': 3}
and observe the touchpad matrix change to 3x3.

Change-Id: I0acd76d630710904b0bec9eff1de10de6fec0ec6
Reviewed-on: https://gerrit.chromium.org/gerrit/63436
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>