Minijack: Remove dependency on Django ORM
Rewrite the database used in Minijack backend, so that it can be used on
frontend too. The rewritten db support QuerySet that acts like what
Django support, but with some difference and additional features.

Also, move manage.py up and add minijack_common.py, so that the import
path is relative to the directory minijack. This is for easier deploy
frontend without need of full cros.factory.minijack import path. Add the
path to Makefile, so `make lint` can find the correct import path.

BUG=None
TEST=Start minijack frontend, check every page does works as before.

Change-Id: I29bec53ee3be20e03517c9a14bfbdb97aa882b4e
Reviewed-on: https://chromium-review.googlesource.com/65817
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>