Minijack: Rewrite FileScanner and don't use EventLogWatcher
EventLogWatcher runs as a thread, which causes many weird problems. And
it also shares with Goofy. Some features are not necessary in Minijack.

This change is to rewrite it as FileScanner, which runs as a separated
process and fits the Minijack Worker interface.

BUG=chrome-os-partner:19896
TEST=manual
Run Minijack to parse the log and work well.
Use kill to terminal Minijack gracefully.

Change-Id: Ia69e5f8b26b20b019e21ffc55351dbfb60fa4815
Reviewed-on: https://chromium-review.googlesource.com/167834
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>