Minijack: Move the database check in the connect method
The database is used in several ways:
 (1) Minijack process -> db/sqlite;
 (2) Local Minijack Frontend -> db/sqlite;
 (3) App Engine Minijack Frontend ->
     See the settings.BACKEND_DATABASE:
     -- 'bigquery' -> db/bigquery
     -- 'cloud sql' -> db/cloud_sql

The (1) and (2) needs the path of the Sqlite DB. It is specified in
different ways:
 (1) passed by an argument, --minijack_db;
 (2) search for a list of default location.

We have a check for (2) to check the database existence. An error will
be raised when no database found.

We original used an environ variable, MINIJACK_PROCESS, to info it is
(1), but not (2). However, it made some unittests failed as they miss
this environ variable. We can add this environ variable to each unittest,
but this approach is ugly.

This change is to fix this issue. We ignore the database existence
check in the settings module, and do the check when connecting to the
database. So all the unittests now pass without database-not-found
error, while the Minijack Frontend still has this check.

BUG=chrome-os-partner:22599
TEST=Manual
make test; make lint
Run Minijack process, work well.
Run Minijack Frontend locally, work well when database existed.
Run Minijack Frontend locally, raise the exception when database not existed.

Change-Id: I710038281b61abb1071c2999ab9522a54ba9a71a
Reviewed-on: https://chromium-review.googlesource.com/168694
Reviewed-by: Dean Liao <deanliao@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>