touchscreen_calibration: move the test to public
Since Link has been public, we move the factory_TouchscreenCalibration
test to the factory repo. Note that the code is not changed except
the filename.

We will fix the code style in a subsequent patch; so we add this
test to LINT_BLACKLIST temporarily.

BUG=chrome-os-partner:22981
TEST=None.

Change-Id: I41b8cc33538836ae0db01ec8be26876e138ec6eb
Reviewed-on: https://chromium-review.googlesource.com/170937
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>