touchscreen_calibration: fix the code style meet factory pylint
This patch fixes the pylint problem, primarily the indentation
is changed from 4 spaces to 2 spaces, change the function names
to use Camel Cases, and add comments and docstrings, etc.

It uses the unittest framework instead of the autotest framework.

It also removes some modules that are not used, and add declarations
of class object members in setUp().

As the pylint problems are fixed, it is removed from LINT_BLACKLIST.

BUG=chrome-os-partner:22981
TEST=It should pass make lint.
(cr) $ make lint LINT_WHITELIST="py/test/pytests/touchscreen_calibration.py"

Change-Id: Ia824ef5c533602d17a6d0fedeee42d7592a00122
Reviewed-on: https://chromium-review.googlesource.com/170938
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>