display_point: ignore non numeric input
For the display point test it is possible to put in invalid
values, we should ignore those and only process values that
make sense to avoid getting back NaN. For example we should not
fail if the operator puts in 'abcd', rather we should ignore it
and wait for a valid value such as '2'.

BUG=None
TEST=Run display point test, type in non numeric values and hit enter.
Verify that the input box clears but the test still waits for input.

Change-Id: I23686c2dc2724689d2b68e92a890ce9f47c8eca0
Reviewed-on: https://chromium-review.googlesource.com/170510
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Mohammed Habibulla <moch@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>