Hook factory task instance to test ui.
When we click the "Mark failed" icon on the test ui, the test fails to
stop if it's composed of FactoryTask instances. Hooking the factory task
instance to the test ui, and let test ui call the clean up method on the
task solves the issue.

BUG=None
TEST=make lint test
TEST=manually test on my DUT

Change-Id: Id547396a3036441b0b5d1cac6eb68c8727835e69
Reviewed-on: https://chromium-review.googlesource.com/171548
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>