serial_utils: open a port only when it is not open yet
When using serial.Serial(), the port is immediately open when the
object is created. On the other hand, when the port is None at
object creation, a successive open() is needed to open the port.

In this patch, the open() is invoked only when the port is not
open yet at object creation time.

BUG=chrome-os-partner:23254
TEST=Manual test with an arduino board.

Change-Id: I626e8a3afb2a93b4967e7c57bc523074c9a74c36
Reviewed-on: https://chromium-review.googlesource.com/173075
Reviewed-by: Dean Liao <deanliao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>