touchscreen_calibration: handle test fixture exceptions
When the touchscreen_calibration test is executed, the test fixture
may be not be in its initial state for some reasons. This patch
handles those potential exceptions including the following situations:
- the power of the test fixture is not turned on,
- the test fixture is in the EMERGENCY_STOP state, or
- the test fixture is in the STOP_UP state.

It shows an alert message and instructs what to do to fix the problems.

BUG=chrome-os-partner:23254
TEST=Manually on DUT.
(1) Do not turn on the power of the test fixture initially.
    Observe that it shows an alert message.
(2) When the probe is moving, trigger the emergency stop by inserting
    something into the test fixture a little bit. Watch out the
    operator's safety. Click the test fixture Refresh button on screen.
    Observe that it shows an alert message.

Change-Id: I058fa6ff7c6be25ef40948ad13115234c678c65c
Reviewed-on: https://chromium-review.googlesource.com/176061
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>