touchscreen_calibration: check existence of mounted media initially
Check the existence of the mounted media before a test is actually
executed. In this way, the operator have the opportunity to insert
a USB dongle to save the test result instead of losing the result.

BUG=chrome-os-partner:23254
TEST=Manually on DUT. Launch the test without inserting a USB dongle.
Observe that it will pops up an alert message. Insert a USB dongle
and restart the test. This time the test should be executed correctly.

Change-Id: I4916b5c5b2802c71d13ba6c742f45be7d82bbe6a
Reviewed-on: https://chromium-review.googlesource.com/176062
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>