touchscreen_calibration: add buttons to drive the probe up/down
When all tests are finished, before turning off the power of the test
fixture, the probe should be placed on the DOWN position for safety
reason.

Each time when the system is powered on, the probe should go to the
UP position automatically. If for any reason the probe does not go
to the UP position, a button is provided to force the probe to go up.

BUG=chrome-os-partner:23254
TEST=Manually on DUT.
(1) Click the button: 关机前制具下位停俥  The probe should go down.
(2) Click the button: 测试前制具上位预备  The probe should go up.

Change-Id: I6d7872a6bcd5d229eb4bc6a38a5a96919f109f33
Reviewed-on: https://chromium-review.googlesource.com/176217
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>