touchscreen_calibration: wait for the expected state until timeout
When the fixture is just powered on and the serial port is reset,
it is possible that the probe is initially resting at the DOWN
position and is on the way of going up to the UP position. In this
situation, use a loop to wait for the probe to reach the UP position
until timeout.

BUG=chrome-os-partner:23254
TEST=Manually on DUT
Place the probe on DOWN position. Restart the touchscreen calibration
test. The probe should go up automaticaly and should not report any
fixture state problem.

Change-Id: Id052a1a5b140146aa913d057f63d1cd2178221e2
Reviewed-on: https://chromium-review.googlesource.com/176218
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>