touchscreen_calibration: pop up an alert when test fixture is not connected
Pop up an alert message when the test fixture USB cable is not
connected to the control host.

BUG=chrome-os-partner:23254
TEST=Manually on DUT
Do not plug the USB cable of the test fixture to the control host. Start
the factory test, and observe that it pops up an alert message. After
connecting the USB cable, click on the test fixture refresh button. It
should work normally.

Change-Id: Ic8b31f359aafb82858b8fd7f50860337dee8a450
Reviewed-on: https://chromium-review.googlesource.com/176219
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>