touchscreen_calibration: adjust motor speed according to probe position
The probe used to move up and down at a constant slow speed. For
efficiency, we would like the probe to move at a faster speed in
the beginning, and slows down when approaching the UP/DOWN end
position for safety.

This was the actual firmware used in the field.

BUG=chrome-os-partner:23132
TEST=Manually on DUT.
Conduct the touchscreen calibration test, and observe that it slows
down when approaching the UP/DOWN end position.

Change-Id: I1e426d92f1229ba5b5cffc797fc6b05880d58a0c
Reviewed-on: https://chromium-review.googlesource.com/199010
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>