ts calibration: verify 3 columns of raw sensor data
In the new test fixture, there are 3 columns of metal fingers
on the probe. In this patch, the column numbers corresponding
to the metal fingers are defined. In addition, the lower and
the upper bounds of the legitimate sensor values are also
defined to determine if the AB panel passes the test.

BUG=chrome-os-partner:29140
TEST=Manually test it on a touchscreen calibration test fixture.

Change-Id: I31e9791918c37214228f04a32b33d9cababe01b5
Reviewed-on: https://chromium-review.googlesource.com/201604
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>