ts calibration: provide a shutdown button on factory GUI
When a tester has completed the test, s/he needs to shutdown the
host so that the current small traveler board could be unplugged
and the next one could be plugged in. This patch provides a
power shutdown button on the factory GUI for convenience.

BUG=chrome-os-partner:29140
TEST=Manually test it on a touchscreen calibration test fixture.

Change-Id: Idb0b90d9ea1e1a8890dbdc37fa78f9f897afad0b
Reviewed-on: https://chromium-review.googlesource.com/201606
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>