sysfs_battery: fail appropriately on battery status
If we set the test list to check if the battery status is good
but the battery status is not found in sysfs, we should fail
instead of hanging in the test trying to .lower() a None.

BUG=None
TEST=Manual, run test on DUT that has no status value in sysfs
with 'verify_battery_health_good': True, verify test fails instead
of hanging.

Change-Id: I2f0ec650f6611061db878b64ad6a6e8c29617951
Reviewed-on: https://chromium-review.googlesource.com/178618
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@chromium.org>
Commit-Queue: Justin Chuang <jchuang@chromium.org>