battery_current: verify the battery exists before testing
We should verify that the battery is actually in place before trying
to get statistics about it. This also fills in the battery path
for the power.py module, without this you could see python errors in
this test where the battery_path value is None.

BUG=None
TEST=Manual, run test on DUT, verify we get a valid path when
running GetChargePct.

Change-Id: I040c593327a9087df9aa7cd25af490f1b8dff707
Reviewed-on: https://chromium-review.googlesource.com/178695
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Justin Chuang <jchuang@chromium.org>