ext_display: Fix DetectDisplayTask to make sure chrome got display info
There is a delay between xrandr seeing external display and chrome
seeing external display.
Before we check display info in VideoTask, we have to make
sure display info is available in DetectDisplayTask.

This CL also relieves the dependency of WaitDisplayThread to
CONNECT = 'connected' and DISCONNECT = 'disconnected' so CONNECT and
DISCONNECT can just represent the task goal, not the exact strings to be
used in the xrandr output pattern.

BUG=chrome-os-partner:24595
TEST=run ext_display test on daisy_spring

Change-Id: I07fc3e885b56f3a2b1179070461bce11e9e4c8f0
Reviewed-on: https://chromium-review.googlesource.com/180391
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>