Start each test case in a separate process.
To support E2E test, we need to be able to create mocks for anything in a
factory test. We also want to be able to interact with the mock object in an
E2E test case. This CL changes invocation.py so that each TestCase of an E2E
test or factory test is invoked in a separate process. With this CL, we can
run the factory test in a thread inside each E2E test case.

Also fix some pylint warnings regarding docstring.

BUG=none
TEST=Run generic test list on my DUT.

Change-Id: I616fd07d72ceb743388e7250e72fc238fc9c8f91
Reviewed-on: https://chromium-review.googlesource.com/180574
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>