Fix the SD card test (using removable_storage) race condition problem
Normally to test an SD card, we request the operator to plug the SD card
with write protection off at the 1st time, and with write protection on
at the 2nd time. But if the operator does not enable write protection at
the 2nd time, the test will still pass now.

That's because in the end of this test, the main thread does not wait
for its child (UI) thread. Thus, one possible case is that python
finished the main thread first after SD card removal, and garbage
collected the UI thread's resources. Since then, the UI thread may not
be able to find some modules and hence can not determine the test has
failed or not. As a consequence, the test always pass.

CQ-DEPEND=CL:193750
BUG=chrome-os-partner:27613
TEST=Manually tested on Big

Change-Id: I25dcab8c9c2a04ae6ce06eaf8a90cef2de51e355
Reviewed-on: https://chromium-review.googlesource.com/193602
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Commit-Queue: Mao Huang <littlecvr@chromium.org>