touchscreen_calibration: use native USB port to monitor its internal state
There are two ports in arduino DUE board. One is the programming port,
and the other is a native USB port. This patch utilizes the native USB
port to monitor the test fixture's main state. Whenever the main state
is changed, it sends the new state to the native USB port.

Note that this patch does not monitor the change of the sensor values
or the motor parameters, etc. We will define a state vector for this
purpose in a subsequent patch.

BUG=chrome-os-partner:28665
TEST=Manually on DUT.
Use screen to monitor the output from arduino by
$ screen -L $PORT 9600
where $PORT is the native programming port, e.g., /dev/ttyACM0

Change-Id: I258eb1509e077892c96c5532bf283b64e2fbd0b3
Reviewed-on: https://chromium-review.googlesource.com/199013
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>