touchscreen_calibration: monitor state vector and refactor with a Fixture class
In a previous patch, the arduino board is able to send its main state
to the native USB port whenever the state is changed. In this patch,
we define a state vector to include the state, the sensor values, and
the motor parameters. We would like the arduino board to send its
state vector whenever any value in the state vector is changed. In
this way, we could monitor everything inside the test fixture.

The arduino program becomes larger now. It is time to encapsulate the
basic functions and the state vector of the fixture to its own class
in a .cpp file, and leave the higher level operations related with
the test procedure in an arduino .ino file, which makes the code more
modular.

BUG=chrome-os-partner:28665
TEST=Manually on DUT.

Change-Id: Iab0b70f70724038974dd74ffda3fdd2c6f20fb99
Reviewed-on: https://chromium-review.googlesource.com/199014
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>