touchscreen_calibration: use PWM to drive the motor
The motor was driven step by step in a loop before. Depending on
the time consumed on an iteration, the period intervals to drive
the motor may not be uniform. We observed at factory that the
motor generated significant vibration.

Using Pulse Width Modulation (PWM) to drive the motor with regular
period intervals is supposed to help improve the problem of the
motor vibration.

One primary difference when driving the motor using PWM lies in that
you need to explicitly stop the motor when it reaches its destination.
Otherwise, the probe keeps going which may lead to some kind of damage.

The probe moves at a faster speed in the beginning, and slows down when
approaching the end position for safety reason.

BUG=chrome-os-partner:28663
TEST=Manually on DUT.
Conduct the touchscreen calibration test, and observe that it drives
the motor smoothly.

Change-Id: I47594a74c75201b01ae5d37f9302ad9725eacb09
Reviewed-on: https://chromium-review.googlesource.com/199011
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>