touchscreen_calibration: determine the initial pwm frequency
If the probe does not park at the DOWN position, use a slow speed.
Otherwise, there is no way to know when to slow down.

If the probe does park at the DOWN position, use a fast speed.
It is faster to move the probe to the UP position in this way.

BUG=chrome-os-partner:28665
TEST=Manually on DUT.

Change-Id: I90c34b7bdda5164d22ca2ec49a67d95450287c27
Reviewed-on: https://chromium-review.googlesource.com/199490
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>