touchscreen_calibration: using timer interrupt to slow down the motor
In this patch, we add a timer interrupt to slow down the motor.
Hence, when either of the following two conditions is met,
the probe slows down.
(1) count is greater than DISTANCE_TO_SLOW_DOWN, or
(2) the elapsed time is greater than TIME_TO_SLOW_DOWN.

Since there is not built-in library for adding a timer interrupt,
we add a 3rd-party library DueTimer for the purpose.

The speed adjustment method using "count" is also moved to the
arduino program since it should be considered as a high-level
operation instead of a basic one of the Fixture class.

BUG=chrome-os-partner:28897
TEST=Manually on DUT.

Change-Id: I77cd15f2490beafd69b2c31294f1c57f2f3fd80e
Reviewed-on: https://chromium-review.googlesource.com/200022
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>