Add ArduinoDigitalPinController class
For some tests, we may need Arduino to switch between something. For
example: the VSWR test may require to switch between 2.4G and 5G if the
module does not support both working at the same time. The task is to
simply pull or push a pin to high or low, thus, adding this class for
convenient use.

BUG=None
TEST=Manually tested

Change-Id: Id7a9ffab6a76c8852db04dbd295c9bd27b1ddad0
Reviewed-on: https://chromium-review.googlesource.com/200020
Reviewed-by: Dean Liao <deanliao@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Commit-Queue: Mao Huang <littlecvr@chromium.org>