ts calibration: use configuration file for sys fs data
When conducting the touchscreen calibration test, we usually need
to write some data to sys fs before and after probing the raw sensor
data. These commands are currently written as statements in the test.
An example is a python statement in the test:

  ...
  self.WriteSysfs('64001C14')
  ...

which changes the touch gain.

This is not scalable in that it is for a particular build only.
Next time, if we have builds for other distinct boards, we need to
overwrite those statements. An ideal method would be to provide a
distinct configuration file for every board. Hence, the above
statement would be some items in the configuration file as

  [PreRead]
  ...
  64001C14: Set touch gain to 20.
  ...

BUG=chrome-os-partner:29672
TEST=Manually test it on a touchscreen calibration test fixture.

Change-Id: I90e9ae777d735c1569de7061df8fee163f69726c
Reviewed-on: https://chromium-review.googlesource.com/204221
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>