Create dummy devices.csv for testing.
When running dummy shopfloor server using factory.par for testing, it
fails to locate devices.csv because testdata are not included in
factory.par. This CL creates a dummy devices.csv file for testing if
the program finds that it's running inside a factory.par.

BUG=chrome-os-partner:26795
CQ-DEPEND=CL:206436
TEST=Manually on my MobLab.

Change-Id: I462b17e035621a260c326eab092758db4f31f623
Reviewed-on: https://chromium-review.googlesource.com/206788
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>