Serial_util: add return multiple ports functionality
For the case of multiple ports with the same driver, and with the same
interface protocol, a new parameter multiple_ports of FindTtyByDriver
is added for detecting all matched ports with given driver and return
in a list.
multiple_ports is False as default to have no influence for original
usage.

_DeviceInterfaceProtocol is moved out from FindTtyByDriver for adding
unittest of interface protocol case.

BUG=chrome-os-partner:29166
TEST=Add unittests for interface_protocol and multiple_ports option

Change-Id: I5c8561d1e992e3640973facbdf7197dc327661e6
Reviewed-on: https://chromium-review.googlesource.com/206805
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Pin-chih Lin <johnylin@chromium.org>