factory-archiver: fix a bug on the reports with .tar.xz format
When assigning .tar.xz in the reports, the compress_in_place is set to
be True because it is reports. However, currently, the .tar.xz haven't
implement the in place compression. This will cause inconsistent of
archiver.metadata and the actual data in archives.

In addition, changing the delimter from "@" to "~" because backend will
use "@" as a sharding annotation.

BUG=chrome-os-partner:27567
TEST=make lint LINT_WHITELIST=" \
  py/lumberjack/archiver.py \
  py/lumberjack/archiver_config.py"
TEST=make test

Change-Id: I40dec43e1ff5c53bf05d19a8f2b256c4db77619b
Reviewed-on: https://chromium-review.googlesource.com/206817
Tested-by: Chun-ta Lin <itspeter@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Chun-ta Lin <itspeter@chromium.org>