Fix temporary source directory permission in toolkit install.
When factory toolkit installs into a test image, it leaves a set of
/tmp/selfgz* temporary source files because these files are owned
by root and the install script fails to remove them at the end.

Change the owner of the source directories back to the original user,
so they can be properly removed.

BUG=chrome-os-partner:30348
TEST=Manually make factory toolkit and install to a test image and
     verify that /tmp/selfgz* are removed.

Change-Id: I67b5a0ba6c6c1581d81cf66644d84a17d88259c4
Reviewed-on: https://chromium-review.googlesource.com/206977
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Jon Salz <jsalz@chromium.org>