factory: Allow console (VT2) customization.
We are having more and more high-resolution devices, and debugging with VT2 on
these devices in factory becomes a problem due to the small fonts in console.

To limit console changes only in factory, we want to do that inside factory init
startup, and only executed when user logins into VT.

The first change includes enabling cursor and setting fonts on large display.
We may add more in future.

BUG=none
TEST=Manually installed on Link and see large fonts :)

Change-Id: Iff9babdc9f8667b6fa958acf79e7ca6cad2ead36
Reviewed-on: https://chromium-review.googlesource.com/207322
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>