Rename goofy_device to goofy_presenter
The "thin host" effort for goofy involves will implement host-based
testing, but with most of the logic remaining on the device, instead
of on the host as was previously planned. This commit exchanges the
host and device sides: goofy now refers to the code running on the
device, while goofy_presenter is the server that runs on the host.

BUG=None
TEST=make test, manual testing

Change-Id: I364d4ade5cde3b523a014222d9017b831ad76ffd
Reviewed-on: https://chromium-review.googlesource.com/208193
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Peter Ammon <pca@chromium.org>
Tested-by: Peter Ammon <pca@chromium.org>