touchscreen_calibration: mitigate the unstable voltage problem
The patch updates the button and sensors status periodically. A
button or a sensor is considered to be pressed or triggered when
the duration is longer than a pre-defined duration threshold.

BUG=chrome-os-partner:31027
TEST=Manually test it on a touchscreen calibration test fixture.
If a user presses the debug button shortly, i.e., less than 0.5
second, the probe will not move. If the user presses the debug
button longer than 0.5 second, the probe will move down or up
as before. The safety sensor is set a much shorter duration bound,
i.e., 0.1 second for the safety reason.

Change-Id: I18b5cfbc3d14a0fd89679be2ceb92b91f542f700
Reviewed-on: https://chromium-review.googlesource.com/210462
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>